Extract `WrapPoint3d` and `WrapPointPad` to a distinct file
* Create a new file `wrap_shape0d.cpp` to house wrapping code for 0-D geometry
  entity.
* Move the C++ wrapper classes `WrapPoint3d` and `WrapPointPad` from
  file `wrap_World.cpp` to `wrap_shape0d.cpp`.

Claude code prompt: Extract WrapPoint3d and WrapPointPad from wrap_World.cpp to
a new file wrap_shape0d.cpp
