+++ cpp/modmesh/universe/CMakeLists.txt
@@ -24,6 +24,7 @@ set(MODMESH_UNIVERSE_PYMODHEADERS
 set(MODMESH_UNIVERSE_PYMODSOURCES
     ${CMAKE_CURRENT_SOURCE_DIR}/pymod/universe_pymod.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/pymod/wrap_bernstein.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/pymod/wrap_shape0d.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/pymod/wrap_World.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/pymod/wrap_polygon.cpp
     CACHE FILEPATH "" FORCE)
